requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
networkx>=3.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
streamlit>=1.20.0

python-Levenshtein>=0.20.0
